moving things around, minor fixes
