+++ Cargo.toml
@@ -2,9 +2,10 @@
 
 members = [
     "common/rust/ast/core",
-    "common/rust/file-manager-client",
-    "common/rust/json-rpc",
     "common/rust/ast/macros",
+    "common/rust/file-manager",
+    "common/rust/ide",
+    "common/rust/json-rpc",
     "common/rust/macro-utils",
     "common/rust/parser",
     "common/rust/prelude",
