+++ common/rust/ide/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "file-manager-client"
+name = "ide"
 version = "0.1.0"
 authors = ["Enso Team <contact@luna-lang.org>"]
 edition = "2018"
@@ -10,6 +10,7 @@ crate-type = ["cdylib", "rlib"]
 [dependencies]
 json-rpc     = { version = "0.1.0"  , path = "../json-rpc" }
 prelude      = { version = "0.1.0"  , path = "../prelude" }
+file-manager-client = { version = "0.1.0"  , path = "../file-manager" }
 serde        = { version = "1.0", features = ["derive"] }
 serde_json   = "1.0"
 shrinkwraprs = "0.3.0"
