+++ common/rust/shapely/core/src/lib.rs
@@ -41,6 +41,7 @@ where G: Generator<Return = ()> + Unpin {
 pub struct EmptyIterator<T>(PhantomData<T>);
 
 impl<T> EmptyIterator<T> {
+    /// Create a new empty iterator.
     pub fn new() -> Self {
         Self(PhantomData)
     }
