--- common/rust/transport/Cargo.toml
@@ -1,38 +1,0 @@
-[package]
-name = "file-manager-client"
-version = "0.1.0"
-authors = ["Enso Team <contact@luna-lang.org>"]
-edition = "2018"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-json-rpc     = { version = "0.1.0"  , path = "../json-rpc" }
-prelude      = { version = "0.1.0"  , path = "../prelude" }
-serde        = { version = "1.0", features = ["derive"] }
-serde_json   = "1.0"
-shrinkwraprs = "0.3.0"
-futures      = "0.3.1"
-wasm-bindgen = "0.2.58"
-js-sys       = "0.3.35"
-failure      = "0.1.6"
-console_error_panic_hook = "0.1.6"
-paste        = "0.1.6"
-uuid = { version = "0.8", features = ["serde", "v5"] }
-
-[dependencies.web-sys]
-version = "0.3.22"
-features = [
-  'Blob',
-  'console',
-  'CloseEvent',
-  'Document',
-  'Element',
-  "ErrorEvent",
-  "MessageEvent",
-  'HtmlElement',
-  'Node',
-  'WebSocket',
-  'Window',
-]
